RL-18: moved styles to separate files
