+++ client/src/components/AuthBtn/styles.ts
@@ -1,0 +1,9 @@
+import { createStyles, makeStyles } from '@material-ui/core'
+
+export const useStyles = makeStyles(() =>
+    createStyles({
+        button: {
+            whiteSpace: 'nowrap',
+        },
+    })
+)
