+++ client/src/components/MessageSnackbar/MessageSnackbar.tsx
@@ -1,21 +1,13 @@
 import React, { FC, useEffect, useState } from 'react'
 import Snackbar from '@material-ui/core/Snackbar'
-import { Alert } from '@material-ui/lab'
-import { createStyles, makeStyles } from '@material-ui/core'
+import Alert from '@material-ui/lab/Alert'
 import { useTypedSelector } from '../../hooks/useTypedSelector'
 import { useActions } from '../../hooks/useActions'
 import { selectSnackbarData } from '../../store/reducers/selectors'
-
-const useStyles = makeStyles(() =>
-    createStyles({
-        alignItem: {
-            alignItems: 'center',
-        },
-    })
-)
+import { useStyles } from './styles'
 
 const MessageSnackbar: FC = () => {
-    const classes = useStyles()
+    const { alignItem } = useStyles()
     const [open, setOpen] = useState(false)
     const { setError, setMessage } = useActions()
 
@@ -47,7 +39,7 @@ const MessageSnackbar: FC = () => {
         >
             <Alert
                 severity={snackbarData.severityError ? 'error' : 'success'}
-                className={classes.alignItem}
+                className={alignItem}
             >
                 {snackbarData.message}
             </Alert>
