+++ client/src/components/SearchForm/styles.ts
@@ -1,0 +1,19 @@
+import { createStyles, makeStyles } from '@material-ui/core'
+
+export const useStyles = makeStyles((theme) =>
+    createStyles({
+        searchForm: {
+            display: 'flex',
+            alignItems: 'center',
+            [theme.breakpoints.down('sm')]: {
+                flexWrap: 'wrap',
+                justifyContent: 'center',
+                width: '100%',
+            },
+        },
+        buttonsContainer: {
+            display: 'flex',
+            justifyContent: 'space-between',
+        },
+    })
+)
