+++ client/src/components/SearchInput/SearchInput.tsx
@@ -1,39 +1,13 @@
 import React, { FC } from 'react'
-import { createStyles, makeStyles, Theme } from '@material-ui/core/styles'
 import TextField from '@material-ui/core/TextField'
 import SearchIcon from '@material-ui/icons/Search'
-import { IconButton } from '@material-ui/core'
+import IconButton from '@material-ui/core/IconButton'
 import { useFormik } from 'formik'
 import { searchValidationSchema } from './validationSheme'
 import { useActions } from '../../hooks/useActions'
 import { useTypedSelector } from '../../hooks/useTypedSelector'
 import { selectStatusZipCodeApi } from '../../store/reducers/settings/selectors'
-
-const useStyles = makeStyles((theme: Theme) =>
-    createStyles({
-        form: {
-            display: 'flex',
-            alignItems: 'center',
-        },
-        input: {
-            margin: theme.spacing(1),
-            width: '40ch',
-            [theme.breakpoints.down('sm')]: {
-                order: -1,
-            },
-        },
-        searchBtn: {
-            height: 50,
-            border: '1px solid rgba(0, 0, 0, 0.23)',
-            borderRadius: theme.shape.borderRadius,
-            backgroundColor: theme.palette.whiteBackground.main,
-            marginRight: 12,
-        },
-        icon: {
-            fontSize: theme.spacing(3),
-        },
-    })
-)
+import { useStyles } from './styles'
 
 const SearchInput: FC = () => {
     const { form, input, searchBtn, icon } = useStyles()
