+++ client/src/components/WeatherCard/WeatherCard.tsx
@@ -1,54 +1,23 @@
 import React, { FC, ReactElement, useState } from 'react'
-import { makeStyles } from '@material-ui/core/styles'
 import Card from '@material-ui/core/Card'
-import { createStyles } from '@material-ui/core'
 import Grid from '@material-ui/core/Grid'
 import IconButton from '@material-ui/core/IconButton'
 import Close from '@material-ui/icons/Close'
 import TableWeather from '../TableWeather/TableWeather'
 import WeatherCardSettings from '../WeatherCardSettings/WeatherCardSettings'
-import { useActions } from '../../hooks/useActions'
+import InteractiveMap from '../InteractiveMap/InteractiveMap'
 import GraphWeather from '../GraphWeather/GraphWeather'
 import { useTypedSelector } from '../../hooks/useTypedSelector'
+import { useActions } from '../../hooks/useActions'
+import { MAP_BTN_TEXT, TABLE_BTN_TEXT } from '../../config/constants'
+import { selectShownOnGraphLocations } from '../../store/reducers/settings/selectors'
+import { useStyles } from './styles'
 import {
     selectFavouritesForecastDataInUnits,
     selectTimezone,
     selectWeatherForecastInUnits,
     selectWeatherUnits,
 } from '../../store/reducers/weather/selectors'
-import InteractiveMap from '../InteractiveMap/InteractiveMap'
-import { MAP_BTN_TEXT, TABLE_BTN_TEXT } from '../../config/constants'
-import { selectShownOnGraphLocations } from '../../store/reducers/settings/selectors'
-
-export const useStylesCard = makeStyles((theme) =>
-    createStyles({
-        card: {
-            position: 'relative',
-            minWidth: 300,
-            height: 285,
-            borderRadius: 40,
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            backgroundColor: theme.palette.secondary.main,
-        },
-        table: {
-            paddingLeft: 20,
-        },
-        settings: {
-            display: 'flex',
-            flexDirection: 'column',
-            alignItems: 'center',
-            justifyContent: 'space-around',
-            paddingRight: 10,
-        },
-        icon: {
-            position: 'absolute',
-            top: 15,
-            right: 20,
-        },
-    })
-)
 
 interface IWeatherCardProps {
     id: string
@@ -56,7 +25,7 @@ interface IWeatherCardProps {
 
 export const WeatherCard: FC<IWeatherCardProps> = ({ id }) => {
     const [viewState, changeViewState] = useState(TABLE_BTN_TEXT)
-    const classes = useStylesCard()
+    const { card, table, settings, icon } = useStyles()
     const forecast = useTypedSelector(selectWeatherForecastInUnits)
     const timezone = useTypedSelector(selectTimezone)
     const unitsDegrees = useTypedSelector(selectWeatherUnits)
@@ -100,19 +69,19 @@ export const WeatherCard: FC<IWeatherCardProps> = ({ id }) => {
     }
 
     return (
-        <Card className={classes.card}>
+        <Card className={card}>
             <Grid container>
-                <Grid className={classes.table} item xs={9} sm={10}>
+                <Grid className={table} item xs={9} sm={10}>
                     {renderContent(viewState)}
                 </Grid>
-                <Grid className={classes.settings} item xs={3} sm={2}>
+                <Grid className={settings} item xs={3} sm={2}>
                     <WeatherCardSettings
                         swapViewMod={swapViewMod}
                         viewState={viewState}
                     />
                 </Grid>
                 <IconButton
-                    className={classes.icon}
+                    className={icon}
                     onClick={deleteCardClick}
                     aria-label="delete"
                 >
