+++ client/src/components/WeatherCardSettings/styles.ts
@@ -1,0 +1,13 @@
+import { createStyles, makeStyles } from '@material-ui/core'
+
+export const useStyles = makeStyles(() =>
+    createStyles({
+        settings: {
+            display: 'flex',
+            height: 200,
+            flexDirection: 'column',
+            alignItems: 'center',
+            justifyContent: 'space-around',
+        },
+    })
+)
