+++ client/src/pages/Home/index.tsx
@@ -1,5 +1,4 @@
 import React, { FC, useEffect } from 'react'
-import { makeStyles, createStyles } from '@material-ui/core'
 import Header from '../../components/Header/Header'
 import Footer from '../../components/Footer/Footer'
 import { useActions } from '../../hooks/useActions'
@@ -9,40 +8,11 @@ import {
     DEGREES_FAHRENHEIT,
     KEY_TOKEN_IN_LOCAL_STORAGE,
     KEY_UNITS_IN_LOCAL_STORAGE,
-    ThemesEnum,
 } from '../../config/constants'
-import BackgroundImage from '../../assets/image/main/world_map.jpg'
-import BackgroundImageDark from '../../assets/image/main/world_map_dark.jpg'
-
-export const useStylesHome = makeStyles((theme) =>
-    createStyles({
-        wrapper: {
-            display: 'flex',
-            flexDirection: 'column',
-            minHeight: '100vh',
-        },
-        content: {
-            maxWidth: 1024,
-            width: '100%',
-            flex: '1 0 0',
-            boxSizing: 'border-box',
-            marginLeft: 'auto',
-            marginRight: 'auto',
-            paddingLeft: 16,
-            paddingRight: 16,
-            paddingBottom: 16,
-            backgroundSize: 'contain',
-            backgroundRepeat: 'no-repeat',
-            backgroundImage:
-                theme.palette.type === ThemesEnum.LIGHT_THEME
-                    ? `url(${BackgroundImage})`
-                    : `url(${BackgroundImageDark})`,
-        },
-    })
-)
+import { useStyles } from './styles'
 
 const Home: FC = () => {
-    const classes = useStylesHome()
+    const { wrapper, content } = useStyles()
     const { auth, getWeatherInCurrentLocation, switchUnits } = useActions()
 
     useEffect(() => {
@@ -56,13 +26,12 @@ const Home: FC = () => {
             switchUnits(DEGREES_FAHRENHEIT)
         }
         getWeatherInCurrentLocation()
-        // eslint-disable-next-line react-hooks/exhaustive-deps
     }, [])
 
     return (
-        <div className={classes.wrapper}>
+        <div className={wrapper}>
             <Header />
-            <main className={classes.content}>
+            <main className={content}>
                 <DragAndDropSpace />
             </main>
             <Footer />
