+++ client/src/service/WeatherService.ts
