+++ client/src/service/types.ts
